Move collate & evaluate to quiz.test
