+++ tests/simple/question.test.js
@@ -1114,172 +1114,3 @@ describe('Delete question', () => {
     });
   });
 });
-describe('Collate questions', () => {
-  describe('Create without login', () => {
-    test('should return 404', async () => {
-      const res = await app.inject({
-        method: 'POST',
-        url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS)
-      });
-
-      expect(res.statusCode).toBe(404);
-      expect(res.body).toBeUndefined();
-    });
-  });
-
-  describe('Create with login', () => {
-    test('should return 404', async () => {
-      let cookies = getLoginCookieFor(userData.updatedFirstUser);
-
-      const res = await app.inject({
-        method: 'POST',
-        url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS),
-        cookies
-      });
-
-      expect(res.statusCode).toBe(404);
-      expect(res.body).toBeUndefined();
-    });
-  });
-
-  describe('Update without login', () => {
-    test('should return 404', async () => {
-      const res = await app.inject({
-        method: 'PATCH',
-        url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS)
-      });
-
-      expect(res.statusCode).toBe(404);
-      expect(res.body).toBeUndefined();
-    });
-  });
-
-  describe('Update with login', () => {
-    test('should return 404', async () => {
-      let cookies = getLoginCookieFor(userData.updatedFirstUser);
-
-      const res = await app.inject({
-        method: 'POST',
-        url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS),
-        cookies
-      });
-
-      expect(res.statusCode).toBe(404);
-      expect(res.body).toBeUndefined();
-    });
-  });
-
-  describe('Delete without login', () => {
-    test('should return 404', async () => {
-      const res = await app.inject({
-        method: 'POST',
-        url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS)
-      });
-
-      expect(res.statusCode).toBe(404);
-      expect(res.body).toBeUndefined();
-    });
-  });
-
-  describe('Delete with login', () => {
-    test('should return 404', async () => {
-      let cookies = getLoginCookieFor(userData.updatedFirstUser);
-
-      const res = await app.inject({
-        method: 'POST',
-        url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS),
-        cookies
-      });
-
-      expect(res.statusCode).toBe(404);
-      expect(res.body).toBeUndefined();
-    });
-  });
-
-  describe('Fetch', async () => {
-    const res = await app.inject({
-      method: 'GET',
-      url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS)
-    });
-
-    test('should return 200', () => {
-      expect(res.statusCode).toBe(200);
-      expect(res.body).toBeDefined();
-      expect(res.body).not.toBeNull();
-    });
-
-    test('should return valid body', () => {
-      let body = JSON.parse(res.body);
-
-      expect(body).toHaveProperty('quiz');
-      expect(body).toHaveProperty('question');
-      expect(body.quiz).not.toBeNull();
-      expect(body.quiz).not.toBe({});
-      expect(Array.isArray(body.question)).toBe(true);
-      let allQnsHaveFormat = body.question.every(
-        (qn) => qn.answer_format !== undefined && qn.answer_format !== null
-      );
-      expect(allQnsHaveFormat).toBe(true);
-      let allQnsHaveNumber = body.question.every(
-        (qn) => qn.answer_format !== undefined && qn.answer_format !== null
-      );
-      expect(allQnsHaveNumber).toBe(true);
-      let allQnsHaveText = body.question.every(
-        (qn) => qn.answer_format !== undefined && qn.answer_format !== null
-      );
-      expect(allQnsHaveText).toBe(true);
-    });
-  });
-
-  describe('Fetch with answer bypass', async () => {
-    const res = await app.inject({
-      method: 'GET',
-      url: endpoints.answersUrl.replace('{threeWords}', THREE_WORDS)
-    });
-
-    test('should return 200', () => {
-      expect(res.statusCode).toBe(200);
-      expect(res.body).toBeDefined();
-      expect(res.body).not.toBeNull();
-    });
-    test('should return valid body', () => {
-      let body = JSON.parse(res.body);
-
-      expect(body).toHaveProperty('quiz');
-      expect(body).toHaveProperty('question');
-      expect(body.quiz).not.toBeNull();
-      expect(body.quiz).not.toBe({});
-      expect(Array.isArray(body.question)).toBe(true);
-      let allQnsHaveFormat = body.question.every(
-        (qn) => qn.answer_format !== undefined && qn.answer_format !== null
-      );
-      expect(allQnsHaveFormat).toBe(true);
-      let allQnsHaveNumber = body.question.every(
-        (qn) => qn.answer_format !== undefined && qn.answer_format !== null
-      );
-      expect(allQnsHaveNumber).toBe(true);
-      let allQnsHaveText = body.question.every(
-        (qn) => qn.answer_format !== undefined && qn.answer_format !== null
-      );
-      expect(allQnsHaveText).toBe(true);
-    });
-
-    test('should not return answers', () => {
-      let body = JSON.parse(res.body);
-
-      expect(body).not.toHaveProperty('answer');
-      expect(body).not.toHaveProperty('answers');
-      let optionsFetched = body.question
-        .filter(
-          ({ answer_format }) =>
-            answer_format === 'mcq' || answer_format === 'msq'
-        )
-        .map(({ options }) => ({ options }))
-        .flat();
-      let someOptHasAnswer = optionsFetched.some(
-        (opt) => opt.is_answer !== undefined
-      );
-      expect(someOptHasAnswer).toBe(false);
-    });
-  });
-});
