+++ tests/simple/quiz.test.js
@@ -8,13 +8,8 @@ const app = require(join(__dirname, '..', '..', 'lib', 'app'));
 // const db = require(join(__dirname, '..', '..', 'lib', 'orm'));
 
 const { endpoints, data } = require('../constants').quiz;
-const {
-  updatedFirstUser,
-  secondUser,
-  thirdUser,
-  userWithoutEmail,
-  endpoints: userEndpoints
-} = require('../constants').user;
+const { data: userData, endpoints: userEndpoints } =
+  require('../constants').user;
 
 let THREE_WORDS, EMPTY_THREE_WORDS;
 
@@ -102,8 +97,7 @@ describe('Create quiz', () => {
 
   describe('Happy path', () => {
     test('should return 201 with data in body', async () => {
-      let { updatedFirstUser } = data;
-      let tempCookie = await getLoginCookieFor(updatedFirstUser); // Login
+      let tempCookie = await getLoginCookieFor(userData.updatedFirstUser); // Login
       let { singleQuiz } = data;
 
       // Create request
@@ -122,7 +116,7 @@ describe('Create quiz', () => {
 
       expect(body.three_words).toBeDefined();
       expect(body.three_words).not.toBeNull();
-      expect(body).toMatchObject(updatedFirstUser);
+      expect(body).toMatchObject(userData.updatedFirstUser);
 
       THREE_WORDS = body.three_words; // setting constant for further tests
     });
@@ -175,7 +169,7 @@ describe('Create quiz', () => {
 
   describe('Send empty data', () => {
     test('should return 201', async () => {
-      let tempCookie = await getLoginCookieFor(secondUser);
+      let tempCookie = await getLoginCookieFor(userData.secondUser);
 
       const res = await app.inject({
         method: 'POST',
@@ -278,7 +272,7 @@ describe('Create impure quiz', () => {
       const log = await app.inject({
         method: 'PATCH',
         url: userEndpoints.loginUser,
-        body: JSON.stringify(updatedFirstUser)
+        body: JSON.stringify(userData.updatedFirstUser)
       });
 
       expect(log.headers).not.toBeNull();
@@ -319,7 +313,7 @@ describe('Update quiz', () => {
   describe('With login', () => {
     test('should return 200 with updated data', async () => {
       let { updatedSingleQuiz } = data;
-      let tempCookie = await getLoginCookieFor(updatedFirstUser);
+      let tempCookie = await getLoginCookieFor(userData.updatedFirstUser);
 
       const res = await app.inject({
         method: 'PATCH',
@@ -343,7 +337,7 @@ describe('Update quiz', () => {
   describe('Change username of quiz', () => {
     test('should return 401 irrespective of login', async () => {
       let { quizWithUsernameBypass } = data;
-      let tempCookie = await getLoginCookieFor(updatedFirstUser);
+      let tempCookie = await getLoginCookieFor(userData.updatedFirstUser);
 
       const res = await app.inject({
         method: 'PATCH',
@@ -360,7 +354,7 @@ describe('Update quiz', () => {
   describe('Move state', () => {
     test('should return 200 with updated data', async () => {
       let { quizWithStateChange } = data;
-      let tempCookie = await getLoginCookieFor(updatedFirstUser);
+      let tempCookie = await getLoginCookieFor(userData.updatedFirstUser);
 
       const res = await app.inject({
         method: 'PATCH',
@@ -397,7 +391,7 @@ describe('Update quiz', () => {
   describe('Change three_words of quiz', () => {
     test('should return 200 with old three_words', async () => {
       let { quizWithThreeWordChange } = data;
-      let tempCookie = await getLoginCookieFor(updatedFirstUser);
+      let tempCookie = await getLoginCookieFor(userData.updatedFirstUser);
 
       const res = await app.inject({
         method: 'PATCH',
@@ -458,7 +452,7 @@ describe('Update quiz', () => {
 
   describe('Send empty data', () => {
     test('should return 400', async () => {
-      let tempCookie = await getLoginCookieFor(secondUser);
+      let tempCookie = await getLoginCookieFor(userData.secondUser);
 
       const res = await app.inject({
         method: 'PATCH',
@@ -517,7 +511,7 @@ describe('Delete quiz', () => {
 
   describe('With login for different user quiz', () => {
     test('should return 401', async () => {
-      let tempCookie = await getLoginCookieFor(secondUser);
+      let tempCookie = await getLoginCookieFor(userData.secondUser);
 
       const res = await app.inject({
         method: 'DELETE',
@@ -532,7 +526,7 @@ describe('Delete quiz', () => {
 
   describe('Happy path', () => {
     test('should return 204', async () => {
-      let tempCookie = await getLoginCookieFor(updatedFirstUser);
+      let tempCookie = await getLoginCookieFor(userData.updatedFirstUser);
 
       const res = await app.inject({
         method: 'DELETE',
@@ -576,7 +570,7 @@ describe('Delete quiz', () => {
 describe('Fetch by username', () => {
   describe('Non-existant user', () => {
     test('should return 404', async () => {
-      let { username } = userWithoutEmail; // This user will never be created
+      let { username } = userData.userWithoutEmail; // This user will never be created
       let url = endpoints.fetchByUser.replace('{username}', username);
 
       const res = await app.inject({ method: 'GET', url });
@@ -588,7 +582,7 @@ describe('Fetch by username', () => {
 
   describe('Existing user without any quiz', () => {
     test('should return 200', async () => {
-      let { username } = thirdUser;
+      let { username } = userData.thirdUser;
 
       const res = await app.inject({
         method: 'GET',
@@ -599,7 +593,7 @@ describe('Fetch by username', () => {
     });
 
     test('should return data valid to schema', async () => {
-      let { username } = thirdUser;
+      let { username } = userData.thirdUser;
 
       const res = await app.inject({
         method: 'GET',
@@ -620,7 +614,7 @@ describe('Fetch by username', () => {
     });
 
     test('should return 0 docs', async () => {
-      let { username } = thirdUser;
+      let { username } = userData.thirdUser;
 
       const res = await app.inject({
         method: 'GET',
@@ -636,7 +630,7 @@ describe('Fetch by username', () => {
 
   describe('Existing user with quizzes', () => {
     test('should return 200', async () => {
-      let { username } = updatedFirstUser;
+      let { username } = userData.updatedFirstUser;
 
       const res = await app.inject({
         method: 'GET',
@@ -647,7 +641,7 @@ describe('Fetch by username', () => {
     });
 
     test('should return data valid to schema', async () => {
-      let { username } = updatedFirstUser;
+      let { username } = userData.updatedFirstUser;
 
       const res = await app.inject({
         method: 'GET',
@@ -668,7 +662,7 @@ describe('Fetch by username', () => {
     });
 
     test('should return more than 0 docs', async () => {
-      let { username } = updatedFirstUser;
+      let { username } = userData.updatedFirstUser;
 
       const res = await app.inject({
         method: 'GET',
@@ -682,7 +676,7 @@ describe('Fetch by username', () => {
     });
 
     test('should return quiz data', async () => {
-      let { username } = updatedFirstUser;
+      let { username } = userData.updatedFirstUser;
 
       const res = await app.inject({
         method: 'GET',
@@ -698,6 +692,203 @@ describe('Fetch by username', () => {
       expect(quizIsInFetchData).toBe(true);
     });
   });
+});
+
+describe('Collate quiz', () => {
+  describe('Create without login', () => {
+    test('should return 404', async () => {
+      const res = await app.inject({
+        method: 'POST',
+        url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS)
+      });
+
+      expect(res.statusCode).toBe(404);
+      expect(res.body).toBeUndefined();
+    });
+  });
+
+  describe('Create with login', () => {
+    test('should return 404', async () => {
+      let cookies = getLoginCookieFor(userData.updatedFirstUser);
+
+      const res = await app.inject({
+        method: 'POST',
+        url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS),
+        cookies
+      });
+
+      expect(res.statusCode).toBe(404);
+      expect(res.body).toBeUndefined();
+    });
+  });
+
+  describe('Update without login', () => {
+    test('should return 404', async () => {
+      const res = await app.inject({
+        method: 'PATCH',
+        url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS)
+      });
+
+      expect(res.statusCode).toBe(404);
+      expect(res.body).toBeUndefined();
+    });
+  });
+
+  describe('Update with login', () => {
+    test('should return 404', async () => {
+      let cookies = getLoginCookieFor(userData.updatedFirstUser);
+
+      const res = await app.inject({
+        method: 'POST',
+        url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS),
+        cookies
+      });
+
+      expect(res.statusCode).toBe(404);
+      expect(res.body).toBeUndefined();
+    });
+  });
+
+  describe('Delete without login', () => {
+    test('should return 404', async () => {
+      const res = await app.inject({
+        method: 'POST',
+        url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS)
+      });
+
+      expect(res.statusCode).toBe(404);
+      expect(res.body).toBeUndefined();
+    });
+  });
+
+  describe('Delete with login', () => {
+    test('should return 404', async () => {
+      let cookies = getLoginCookieFor(userData.updatedFirstUser);
+
+      const res = await app.inject({
+        method: 'POST',
+        url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS),
+        cookies
+      });
+
+      expect(res.statusCode).toBe(404);
+      expect(res.body).toBeUndefined();
+    });
+  });
+
+  describe('Fetch', async () => {
+    const res = await app.inject({
+      method: 'GET',
+      url: endpoints.specificUrl.replace('{threeWords}', THREE_WORDS)
+    });
+
+    test('should return 200', () => {
+      expect(res.statusCode).toBe(200);
+      expect(res.body).toBeDefined();
+      expect(res.body).not.toBeNull();
+    });
+
+    test('should return valid body', () => {
+      let body = JSON.parse(res.body);
+
+      expect(body).toHaveProperty('quiz');
+      expect(body).toHaveProperty('question');
+      expect(body.quiz).not.toBeNull();
+      expect(body.quiz).not.toBe({});
+      expect(Array.isArray(body.question)).toBe(true);
+      let allQnsHaveFormat = body.question.every(
+        (qn) => qn.answer_format !== undefined && qn.answer_format !== null
+      );
+      expect(allQnsHaveFormat).toBe(true);
+      let allQnsHaveNumber = body.question.every(
+        (qn) => qn.answer_format !== undefined && qn.answer_format !== null
+      );
+      expect(allQnsHaveNumber).toBe(true);
+      let allQnsHaveText = body.question.every(
+        (qn) => qn.answer_format !== undefined && qn.answer_format !== null
+      );
+      expect(allQnsHaveText).toBe(true);
+    });
+  });
+
+  describe('Fetch with answer bypass', async () => {
+    const res = await app.inject({
+      method: 'GET',
+      url: endpoints.answersUrl.replace('{threeWords}', THREE_WORDS)
+    });
+
+    test('should return 200', () => {
+      expect(res.statusCode).toBe(200);
+      expect(res.body).toBeDefined();
+      expect(res.body).not.toBeNull();
+    });
+    test('should return valid body', () => {
+      let body = JSON.parse(res.body);
+
+      expect(body).toHaveProperty('quiz');
+      expect(body).toHaveProperty('question');
+      expect(body.quiz).not.toBeNull();
+      expect(body.quiz).not.toBe({});
+      expect(Array.isArray(body.question)).toBe(true);
+      let allQnsHaveFormat = body.question.every(
+        (qn) => qn.answer_format !== undefined && qn.answer_format !== null
+      );
+      expect(allQnsHaveFormat).toBe(true);
+      let allQnsHaveNumber = body.question.every(
+        (qn) => qn.answer_format !== undefined && qn.answer_format !== null
+      );
+      expect(allQnsHaveNumber).toBe(true);
+      let allQnsHaveText = body.question.every(
+        (qn) => qn.answer_format !== undefined && qn.answer_format !== null
+      );
+      expect(allQnsHaveText).toBe(true);
+    });
+
+    test('should not return answers', () => {
+      let body = JSON.parse(res.body);
+
+      expect(body).not.toHaveProperty('answer');
+      expect(body).not.toHaveProperty('answers');
+      let optionsFetched = body.question
+        .filter(
+          ({ answer_format }) =>
+            answer_format === 'mcq' || answer_format === 'msq'
+        )
+        .map(({ options }) => ({ options }))
+        .flat();
+      let someOptHasAnswer = optionsFetched.some(
+        (opt) => opt.is_answer !== undefined
+      );
+      expect(someOptHasAnswer).toBe(false);
+    });
+  });
+});
+
+describe('Evaluate quiz', () => {
+  describe('Create without login', () => {
+    test('should return 401');
+  });
+  describe('Create with login');
+  describe('Update without login', () => {
+    test('should return 401');
+  });
+  describe('Update as different user', () => {
+    test('should return 403');
+  });
+  describe('Update with login');
+  describe('Delete without login', () => {
+    test('should return 404');
+  });
+  describe('Delete with login', () => {
+    test('should return 404');
+  });
+  describe('Fetch without login', () => {
+    test('should return 401');
+  });
+  describe('Fetch as different user', () => {
+    test('should return 403');
+  });
+  describe('Fetch with login');
 });
 
 /*
